requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
tweepy==4.14.0
urllib3==2.1.0
google-auth
google-auth-oauthlib
google-api-python-client

